refactor(evaluation): extract shared utilities from deterministic scorer
Move FlatField, flattenFields, ExtractionOutput, and calculateMetrics
into shared.ts so both deterministic and LLM judge scorers can use them.
